bidfastprobot-v6.2-all-in-one/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
openai
pydantic
